requirements.txt
alembic==1.7.4
click==7.1.2
Flask==1.1.4
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.5
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.2
SQLAlchemy==1.4.26
Werkzeug==1.0.1
zipp==3.6.0
flask-login==0.5.0
WTForms==3.0.1
flask-wtf==1.0.0
wtforms[email]
wtforms-sqlalchemy==0.3
sqlalchemy-utils==0.38.1